tech(ts): EV-85 Upgrade project JS to TS
