+++ .gitignore
@@ -22,3 +22,19 @@ mods
 */mods
 data
 gradle*
+node_modules
+developer.id
+/src/main/resources/view
+rev-manifest.json
+*.mod.json
+**/ts/entcore
+**/ts/**/*.js
+**/ts/*.js
+**/ts/**/*.js.map
+**/ts/*.js.map
+**/template/entcore
+**/temp
+**/dist
+**/.DS_Store
+**/view/*.html
+src/main/resources/public/js/behaviours.js
