+++ package.json
@@ -1,0 +1,50 @@
+{
+  "name": "support",
+  "version": "1.12.0",
+  "description": "Allow users to reporting ENT issues via Support application",
+  "main": "gulpfile.js",
+  "scripts": {
+    "coverage": "./node_modules/.bin/jest --config jest.config.js --coverage",
+    "test": "npm run coverage"
+  },
+  "dependencies": {
+    "entcore": "3.11.0",
+    "awesome-typescript-loader": "3.2.1",
+    "axios": "0.16.2",
+    "core-js": "2.4.1",
+    "entcore-toolkit": "^1.0.1",
+    "gulp": "3.9.1",
+    "gulp-clean": "^0.3.2",
+    "gulp-concat": "^2.6.1",
+    "gulp-rev": "^7.1.2",
+    "gulp-rev-replace": "^0.4.3",
+    "gulp-sass": "^3.2.1",
+    "gulp-sourcemaps": "^2.6.0",
+    "merge2": "^1.1.0",
+    "ts-loader": "^3.2.0",
+    "typescript": "2.4.1",
+    "webpack": "3.1.0",
+    "webpack-stream": "3.2.0",
+    "yargs": "^8.0.2",
+    "json-loader": "^0.5.7"
+  },
+  "devDependencies": {
+    "@types/angular": "^1.6.55",
+    "@types/core-js": "0.9.42",
+    "@types/jest": "^22.2.0",
+    "@types/jquery": "^2.0.34",
+    "gulp-append-prepend": "^1.0.8",
+    "moment": "^2.21.0",
+    "underscore": "^1.8.3"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/OPEN-ENT-NG/support"
+  },
+  "author": "",
+  "license": "AGPL-3.0",
+  "bugs": {
+    "url": "https://github.com/OPEN-ENT-NG/support/issues"
+  },
+  "homepage": "https://github.com/OPEN-ENT-NG/support"
+}
