+++ src/main/resources/public/template/create-ticket.html
@@ -32,7 +32,7 @@
 			<div class="two cell">&nbsp;
 			</div>
 			<div class="ten cell">
-				<light-attachments apps="['workspace']" ng-model="ticket.newAttachments"></light-attachments>
+				<attachments apps="['workspace']" ng-model="ticket.newAttachments"></attachments>
 			</div>
 		</div>
 		<div class="row">
