+++ src/main/resources/public/template/edit-ticket.html
@@ -67,7 +67,7 @@
 		</div>
 
 		<div class="row">
-			<light-attachments apps="['workspace']" ng-model="editedTicket.newAttachments"></light-attachments>
+			<attachments apps="['workspace']" ng-model="editedTicket.newAttachments"></attachments>
 		</div>
 		<div class="row">
 			<div ng-if="!editedTicket.processing">
