+++ src/main/resources/view-src/notify/support/bugtracker-issue-closed.html
