+++ src/main/resources/view-src/notify/support/bugtracker-issue-resolved.html
