+++ src/main/resources/view-src/notify/support/bugtracker-issue-updated.html
