+++ src/main/resources/view-src/notify/support/ticket-created.html
