+++ src/main/resources/view-src/notify/support/ticket-updated.html
