+++ src/main/resources/view-src/support.html
@@ -3,10 +3,9 @@
 <head>
 	<title>{{#i18n}}support.title{{/i18n}}</title>
 	<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-	<script src="{{#infra}}js/loader.js{{/infra}}" type="text/javascript"
-		id="context"></script>
-	<script src="/support/public/js/model.js" type="text/javascript"></script>
-	<script src="/support/public/js/controller.js" type="text/javascript"></script>
+	<script src="/support/public/dist/entcore/ng-app.js" type="text/javascript" id="context"></script>
+	<script src="/support/public/dist/application.js" type="text/javascript"></script>
+
 	<link href="/support/public/css/additional.css" rel="stylesheet">
 	<script type="text/ng-template" id="empty"></script>
 </head>
