+++ tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "declaration": false,
+    "moduleResolution": "node",
+    "sourceMap": true,
+    "noImplicitAny": false,
+    "lib": ["es2015", "dom"],
+    "types": [
+      "jquery",
+      "core-js"
+    ]
+  },
+  "exclude": [
+    "build", "node_modules"
+  ]
+}
